Simplify package structure and improve validation handling
Refactored package structure by removing nested `_5._ercuatri` segments for better readability. Added custom exception `AlreadyUsedEmail` with specific handler. Enhanced DTOs with Lombok annotations for cleaner code. Implemented additional validation for user registration and expanded testing coverage.
