+++ build.gradle
@@ -4,7 +4,7 @@ plugins {
 	id 'io.spring.dependency-management' version '1.1.7'
 }
 
-group = 'unq.dapp.2025.1ercuatri.grupoj'
+group = 'unq.dapp.grupoj'
 version = '0.0.1-SNAPSHOT'
 
 java {
@@ -15,6 +15,10 @@ java {
 
 repositories {
 	mavenCentral()
+}
+
+test {
+	useJUnitPlatform()
 }
 
 dependencies {
@@ -37,4 +41,8 @@ dependencies {
 	testAnnotationProcessor 'org.projectlombok:lombok:1.18.34'
 
 	runtimeOnly 'com.mysql:mysql-connector-j'
+
+	//TESTING
+	testImplementation("org.springframework.boot:spring-boot-starter-test:3.4.4")
+
 }
