--- src/main/java/unq/dapp/_5/_ercuatri/grupoj/SoccerGenius/SoccerGeniusApplication.java
@@ -1,13 +1,0 @@
-package unq.dapp._5._ercuatri.grupoj.SoccerGenius;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SoccerGeniusApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SoccerGeniusApplication.class, args);
-	}
-
-}
