--- src/main/java/unq/dapp/_5/_ercuatri/grupoj/SoccerGenius/model/DTO/LoginCredentials.java
@@ -1,15 +1,0 @@
-package unq.dapp._5._ercuatri.grupoj.SoccerGenius.model.DTO;
-
-public class LoginCredentials {
-    public String email;
-    public String password;
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-}
-
