--- src/main/java/unq/dapp/_5/_ercuatri/grupoj/SoccerGenius/model/DTO/UserDTO.java
@@ -1,7 +1,0 @@
-package unq.dapp._5._ercuatri.grupoj.SoccerGenius.model.DTO;
-
-public class UserDTO {
-    public String firstName;
-    public String lastName;
-    public String email;
-}
