--- src/main/java/unq/dapp/_5/_ercuatri/grupoj/SoccerGenius/services/TeamService.java
@@ -1,11 +1,0 @@
-package unq.dapp._5._ercuatri.grupoj.SoccerGenius.services;
-
-import unq.dapp._5._ercuatri.grupoj.SoccerGenius.model.Player;
-import unq.dapp._5._ercuatri.grupoj.SoccerGenius.model.Team;
-
-import java.util.List;
-
-public interface TeamService {
-
-    public List<Player> getTeamPlayers(String teamName);
-}
