--- src/main/java/unq/dapp/_5/_ercuatri/grupoj/SoccerGenius/util/GlobalExceptionHandler.java
@@ -1,41 +1,0 @@
-package unq.dapp._5._ercuatri.grupoj.SoccerGenius.util;
-
-import org.springframework.boot.context.config.ConfigDataResourceNotFoundException;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.util.HashMap;
-import java.util.Map;
-
-//CHECk
-@RestControllerAdvice
-public class GlobalExceptionHandler {
-    @ExceptionHandler(ConfigDataResourceNotFoundException.class)
-    public ResponseEntity<String> handleResourceNotFound(ConfigDataResourceNotFoundException ex) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(ex.getMessage());
-    }
-
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity<Map<String, String>> handleValidationExceptions(MethodArgumentNotValidException ex) {
-        Map<String, String> errors = new HashMap<>();
-
-        ex.getBindingResult().getFieldErrors().forEach(error ->
-                errors.put(error.getField(), error.getDefaultMessage())
-        );
-
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(errors);
-    }
-
-    @ExceptionHandler({org.springframework.http.converter.HttpMessageNotReadableException.class})
-    public ResponseEntity<String> notRedeableHttp() {
-        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Invalid body format");
-    }
-
-    @ExceptionHandler(RuntimeException.class)
-    public ResponseEntity<String> handleGenericRuntime(RuntimeException ex) {
-        return ResponseEntity.status(400).body(ex.getMessage());
-    }
-}
