+++ src/main/java/unq/dapp/grupoj/SoccerGenius/SoccerGeniusApplication.java
@@ -1,0 +1,13 @@
+package unq.dapp.grupoj.SoccerGenius;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SoccerGeniusApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SoccerGeniusApplication.class, args);
+	}
+
+}
