+++ src/main/java/unq/dapp/grupoj/SoccerGenius/model/DTO/LoginCredentials.java
@@ -1,0 +1,12 @@
+package unq.dapp.grupoj.SoccerGenius.model.DTO;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class LoginCredentials {
+    public String email;
+    public String password;
+    }
+
