+++ src/main/java/unq/dapp/grupoj/SoccerGenius/model/DTO/UserDTO.java
@@ -1,0 +1,13 @@
+package unq.dapp.grupoj.SoccerGenius.model.DTO;
+
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class UserDTO {
+    private String firstName;
+    private String lastName;
+    private String email;
+}
