--- src/test/java/unq/dapp/_5/_ercuatri/grupoj/SoccerGenius/SoccerGeniusApplicationTests.java
@@ -1,13 +1,0 @@
-package unq.dapp._5._ercuatri.grupoj.SoccerGenius;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SoccerGeniusApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
