+++ src/test/java/unq/dapp/grupoj/SoccerGenius/SoccerGeniusApplicationTests.java
@@ -1,0 +1,21 @@
+package unq.dapp.grupoj.SoccerGenius;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.ApplicationContext;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+@SpringBootTest
+public class SoccerGeniusApplicationTests {
+
+    @Autowired
+    private ApplicationContext applicationContext;
+
+    @Test
+    void contextLoads() {
+        assertThat(applicationContext).isNotNull();
+    }
+
+}
